dataDump

+++ bin/util/cr-full-dump.sh
@@ -8,8 +8,11 @@
 # This is highly redundant, but can be helpful for those that "just want the data"
 # and don't want to crawl the VoID dataDumps to get it.
 
-see="https://github.com/timrdf/csv2rdf4lod-automation/wiki/CSV2RDF4LOD-not-set"
-CSV2RDF4LOD_HOME=${CSV2RDF4LOD_HOME:?"not set; source csv2rdf4lod/source-me.sh or see $see"}
+#see="https://github.com/timrdf/csv2rdf4lod-automation/wiki/CSV2RDF4LOD-not-set"
+#CSV2RDF4LOD_HOME=${CSV2RDF4LOD_HOME:?"not set; source csv2rdf4lod/source-me.sh or see $see"}
+HOME=$(cd ${0%/*/*} && echo ${PWD%/*})
+export CLASSPATH=$CLASSPATH`$HOME/bin/util/cr-situate-classpaths.sh`
+CSV2RDF4LOD_HOME=${CSV2RDF4LOD_HOME:?$HOME}
 
 see="https://github.com/timrdf/csv2rdf4lod-automation/wiki/Aggregating-subsets-of-converted-datasets"
 CSV2RDF4LOD_PUBLISH_OUR_SOURCE_ID=${CSV2RDF4LOD_PUBLISH_OUR_SOURCE_ID:?"not set; see $see"}
@@ -34,7 +37,7 @@ datasetID=`basename $0 | sed 's/.sh$//'`
 versionID='latest' # Doing it every day is a waste of space for this use case. `date +%Y-%b-%d`
 
 cockpit="$sourceID/$datasetID/version/$versionID"
-base=`echo $CSV2RDF4LOD_BASE_URI | perl -pi -e 's|http://||;s/\./-/g;s|/|-|g'`
+base=`echo $CSV2RDF4LOD_BASE_URI | perl -pi -e 's|http://||;s/\./-/g;s|/|-|g'` # e.g. lofd-tw-rpi-edu
 dumpFileLocal=$base.nt.gz
 
 if [[ "$1" == "--help" ]]; then
@@ -124,33 +127,34 @@ if [ "$dryrun" != "true" ]; then
    cat          $cockpit/automatic/$base-uri-node-occurrences.txt | sort -u           > $cockpit/automatic/$base-uri-nodes.txt
 fi
 
+pushd $cockpit &> /dev/null
+   versionedDataset=`cr-dataset-uri.sh --uri`
+   sourceID=`cr-source-id.sh`   # Saved for later
+   datasetID=`cr-dataset-id.sh` # Saved for later
+   versionID=`cr-version-id.sh` # Saved for later
+popd &> /dev/null
+baseURI="${CSV2RDF4LOD_BASE_URI_OVERRIDE:-$CSV2RDF4LOD_BASE_URI}"
+topVoID="${CSV2RDF4LOD_BASE_URI_OVERRIDE:-$CSV2RDF4LOD_BASE_URI}/void"
+
 echo $cockpit/automatic/$base-uri-nodes.ttl
-#if [ "$dryrun" != "true" ]; then
+if [ "$dryrun" != "true" ]; then
    echo "@prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#> ."                                                                  > $cockpit/automatic/$base-uri-nodes.ttl
    echo "@prefix foaf: <http://xmlns.com/foaf/0.1/> ."                                                                            >> $cockpit/automatic/$base-uri-nodes.ttl
    echo "@prefix void: <http://rdfs.org/ns/void#> ."                                                                              >> $cockpit/automatic/$base-uri-nodes.ttl
    echo "@prefix prov: <http://www.w3.org/ns/prov#> ."                                                                            >> $cockpit/automatic/$base-uri-nodes.ttl
-   echo                                                                                                                           >> $cockpit/automatic/$base-uri-nodes.ttl
-   pushd $cockpit &> /dev/null
-      versionedDataset=`cr-dataset-uri.sh --uri`
-   popd &> /dev/null
-   baseURI="${CSV2RDF4LOD_BASE_URI_OVERRIDE:-$CSV2RDF4LOD_BASE_URI}"
-   topVoID="${CSV2RDF4LOD_BASE_URI_OVERRIDE:-$CSV2RDF4LOD_BASE_URI}/void"
-   cat $cockpit/automatic/$base-uri-nodes.txt | awk -v dataset=$versionedDataset '{print $1,"void:inDataset <"dataset"> ."}'      >> $cockpit/automatic/$base-uri-nodes.ttl
    echo "#3> <> prov:wasAttributedTo [ foaf:name \"`basename $0`\" ]; ."                                                          >> $cockpit/automatic/$base-uri-nodes.ttl
+   echo                                                                                                                           >> $cockpit/automatic/$base-uri-nodes.ttl
+   cat $cockpit/automatic/$base-uri-nodes.txt | awk -v dataset=$versionedDataset '{print $1,"void:inDataset <"dataset"> ."}'      >> $cockpit/automatic/$base-uri-nodes.ttl
    echo "<$topVoID> void:rootResource <$topVoID> ."                                                                               >> $cockpit/automatic/$base-uri-nodes.ttl
    echo "<$topVoID> void:dataDump     <$baseURI/source/$sourceID/file/$datasetID/version/$versionID/conversion/$dumpFileLocal> ." >> $cockpit/automatic/$base-uri-nodes.ttl
-#fi
+fi
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
 
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-#if [ "$dryrun" != "true" ]; then
+if [ "$dryrun" != "true" ]; then
    pushd $cockpit &> /dev/null
       cr-pwd.sh
-      sourceID=`cr-source-id.sh`   # Saved for later
-      datasetID=`cr-dataset-id.sh` # Saved for later
-      versionID=`cr-version-id.sh` # Saved for later
       aggregate-source-rdf.sh automatic/$base-uri-nodes.ttl
    popd &> /dev/null
 
@@ -171,28 +175,40 @@ echo $cockpit/automatic/$base-uri-nodes.ttl
    #                        ________""""""""_____________________""""""____________"""""""""______""""""""""""_________________________
    #                        /var/www/source/healthdata-tw-rpi-edu/file/cr-full-dump/version/latest/conversion/purl-org-twc-health.nt.gz
 
-   # NOTE: this is repeated from bin/aggregate-source-rdf.sh - be sure to align with it.
-   # (update: This might have been superceded by bin/aggregate-source-rdf.sh, check!)
-   sudo="sudo"
-   if [[ `whoami` == root ]]; then
-      sudo=""
-   elif [[ "`stat --format=%U "$CSV2RDF4LOD_PUBLISH_VARWWW_ROOT/source"`" == `whoami` ]]; then
-      sudo=""
-   fi
 
-   symbolic=""
-   wd=""
-   if [[ "$CSV2RDF4LOD_PUBLISH_VARWWW_LINK_TYPE" == "soft" ]]; then
-     symbolic="-sf "
-     wd=`pwd`/
-   fi
+         # NOTE: this is repeated from bin/aggregate-source-rdf.sh - be sure to align with it.
+         # (update: This might have been superceded by bin/aggregate-source-rdf.sh, check!)
+         # (update 24 Apr 2013 - this is superceded by cr-ln-to-www-root.sh publish/lofd-tw-rpi-edu.nt.gz)
+         #sudo="sudo"
+         #if [[ `whoami` == root ]]; then
+         #   sudo=""
+         #elif [[ "`stat --format=%U "$CSV2RDF4LOD_PUBLISH_VARWWW_ROOT/source"`" == `whoami` ]]; then
+         #   sudo=""
+         #fi
+         #
+         #symbolic=""
+         #wd=""
+         #if [[ "$CSV2RDF4LOD_PUBLISH_VARWWW_LINK_TYPE" == "soft" ]]; then
+         #  symbolic="-sf "
+         #  wd=`pwd`/
+         #fi
+         #
+         #wwwFile="$CSV2RDF4LOD_PUBLISH_VARWWW_ROOT/source/$sourceID/file/$datasetID/version/$versionID/conversion/$dumpFileLocal"
+         #echo "$wwwFile"
+         #$sudo rm -f $wwwFile
+         #echo $sudo ln $symbolic "${wd}$cockpit/publish/$dumpFileLocal" $wwwFile
+         #     $sudo ln $symbolic "${wd}$cockpit/publish/$dumpFileLocal" $wwwFile
 
-   wwwFile="$CSV2RDF4LOD_PUBLISH_VARWWW_ROOT/source/$sourceID/file/$datasetID/version/$versionID/conversion/$dumpFileLocal"
-   echo "$wwwFile"
-   $sudo rm -f $wwwFile
-   echo $sudo ln $symbolic "${wd}$cockpit/publish/$dumpFileLocal" $wwwFile
-        $sudo ln $symbolic "${wd}$cockpit/publish/$dumpFileLocal" $wwwFile
-#fi
+   pushd $cockpit &> /dev/null
+      # Replaces duplication above:
+      cr-ln-to-www-root.sh publish/lofd-tw-rpi-edu.nt.gz
+      one_click_dump=`cr-ln-to-www-root.sh -n --url-of-filepath publish/lofd-tw-rpi-edu.nt.gz`
+
+      # In case the triples we snuck in didn't get published into /var/www
+      cr-ln-to-www-root.sh publish/$base.void.ttl
+   popd &> /dev/null
+
+fi
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
 dryrun.sh $dryrun ending
